requirements.txt
streamlit==1.38.0
plotly==5.24.0
anthropic==0.34.0
python-dotenv==1.0.1